Loading message and "loading message"
On the kanban view, when the user logs in or out, instead of seeing the username it says "loading...", i changed the message to show a spinner, that is more design thoughtful.

Also on the login, the mintral logo has a href to flowbite's page, it got changed to the mintral page (this change was added to the navbarBrand in the "not found" page).